correcion y modularizacion handlers y controllers
