+++ api/src/controllers/dietControllers/dietsControllers.js
@@ -1,3 +1,4 @@
+const {Diet } = require('../../db');
 const getDietsController = async () => {
     //console.log('getDietsController')
     const allDiets=await Diet.findAll();
