+++ api/src/routes/foodRouter.js
@@ -1,10 +1,9 @@
 const {Router} = require('express');
-const {getAllFoodHandler, getFoodByNameHandler, postFoodHandler,putFoodHandler} = require('../handlers/foodHandlers');
+const { getFoodHandler,postFoodHandler,putFoodHandler} = require('../handlers/foodHandlers');
 const foodRouter = Router();
 
 
-foodRouter.get('/', getAllFoodHandler );
-foodRouter.get('/', getFoodByNameHandler);
+foodRouter.get('/', getFoodHandler );
 foodRouter.post('/', postFoodHandler );
 foodRouter.put('/', putFoodHandler );
